added node-name-slugifier to centralice additional node name replacer (#110)
