+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 CHANGELOG for Sulu Document Manager
 ===================================
 
+* dev-develop
+    * ENHANCEMENT #110 Added node-name-slugifier to centralice additional node name replacer
+
 * 0.9.1 (2017-03-16)
     * ENHANCEMENT #107 Added VersionNotFoundException
